red pencil is true for percentages off between 5 and 30 %

+++ RedPencil.Entity/Product.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal OriginalPrice { get; set; }
-        public decimal CurrentPrice { get; set; }
+        public double OriginalPrice { get; set; }
+        public double CurrentPrice { get; set; }
         public bool IsRedPencil { get; set; }
     }
 }
